requirements.txt
# Core dependencies
torch>=2.0.0
deepxde>=1.9.0
numpy>=1.24.0
scipy>=1.11.0

# Database
sqlalchemy>=2.0.0

# Web UI
streamlit>=1.28.0

# Data processing and visualization
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.14.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-benchmark>=4.0.0
hypothesis>=6.82.0

# Utilities
pyyaml>=6.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
